use floats

+++ Engine/OLSRegression.cs
@@ -7,9 +7,9 @@ namespace Engine
 {
     internal class OLSRegression
     {
-        internal double R2 { get; private set; }
-        internal double A { get; private set; }
-        internal double B { get; private set; }
+        internal float R2 { get; private set; }
+        internal float A { get; private set; }
+        internal float B { get; private set; }
 
         internal OLSRegression()
         {
@@ -18,12 +18,12 @@ namespace Engine
 
         public void Perform(FullLine[] data)
         {
-            double n = data.Length;
-            double Sx = data.Sum(x => x.m);
-            double Sy = data.Sum(x => x.c);
-            double Sxx = data.Sum(x => x.m * x.m);
-            double Syy = data.Sum(x => x.c * x.c);
-            double Sxy = data.Sum(x => x.m * x.c);
+            float n = data.Length;
+            float Sx = data.Sum(x => x.m);
+            float Sy = data.Sum(x => x.c);
+            float Sxx = data.Sum(x => x.m * x.m);
+            float Syy = data.Sum(x => x.c * x.c);
+            float Sxy = data.Sum(x => x.m * x.c);
 
             B = ((n * Sxy) - (Sx * Sy)) /
                 ((n * Sxx) - (Sx * Sx));
